pyproject.toml
[project]
name = "scrap"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Virtualstar", email = "Virtualstar0125@gmail.com" }
]
dependencies = [
    "scrapybara>=2.4.7",
    "undetected-playwright-patch>=1.40.0.post1700587210000",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/scrap"]
